[sharktank] Fix attention dtype (#1243)
Fixes bug in refactor (https://github.com/nod-ai/shark-ai/pull/1098)
that removed specifying different dtypes for cache vs attention.

---------

+++ sharktank/sharktank/export_layer/export_kv_cache.py
@@ -65,7 +65,8 @@ def main():
         attn_head_count=attn_head_count,
         attn_head_dim=attn_head_dim,
         shard_count=args.sharding,
-        dtype=torch.float32,
+        cache_dtype=torch.float32,
+        attn_dtype=torch.float32,
         device=None,
     )
